+++ src/main/java/BACK/Dao/DaoImpl/WishlistDaoImpl.java
@@ -86,7 +86,7 @@ public class WishlistDaoImpl implements WDao {
                             rs.getString("name"),
                             rs.getString("type"),
                             rs.getInt("score"),
-                            rs.getDouble("price"),
+                            rs.getBigDecimal("price"),
                             rs.getString("overview")
                     );
                     games.add(game);
